+++ examples/src/examples/thingie.clj
@@ -1,13 +1,15 @@
 (ns examples.thingie
   (:require [ring.util.http-response :refer :all]
             [compojure.api.sweet :refer :all]
-            [ring.swagger.schema :refer [defmodel]]
             [schema.core :as s]))
 
-(defmodel Total {:total Long})
-(defmodel Thingie {:id Long
-                   :hot Boolean
-                   :tag (s/enum :kikka :kukka)})
+(s/defschema Total {:total Long})
+(s/defschema Thingie {:id Long
+                      :hot Boolean
+                      :tag (s/enum :kikka :kukka)
+                      :chief [{:name String
+                               :type #{{:id String}}}]})
+(s/defschema FlatThingie (dissoc Thingie :chief))
 
 (defroutes* legacy-route
   (GET* "/legacy/:value" [value]
@@ -22,33 +24,38 @@
     (context "/api" []
 
       (GET* "/plus" []
-        :return Total
+        :return       Total
         :query-params [x :- Long {y :- Long 1}]
         :summary      "x+y with query-parameters. y defaults to 1."
         (ok {:total (+ x y)}))
 
       (POST* "/minus" []
-        :return Total
-        :body-params  [x :- Long y :- Long]
-        :summary      "x-y with body-parameters."
+        :return      Total
+        :body-params [x :- Long y :- Long]
+        :summary     "x-y with body-parameters."
         (ok {:total (- x y)}))
 
       (GET* "/times/:x/:y" []
-        :return Total
-        :path-params  [x :- Long y :- Long]
-        :summary      "x*y with path-parameters"
+        :return      Total
+        :path-params [x :- Long y :- Long]
+        :summary     "x*y with path-parameters"
         (ok {:total (* x y)}))
 
       legacy-route
 
       (GET* "/echo" []
-        :return   Thingie
-        :query    [thingie Thingie]
-        :summary  "echos a thingie from query-params"
-        (ok thingie)) ;; here be coerced thingie
+        :return   FlatThingie
+        :query    [thingie FlatThingie]
+        :summary  "echos a FlatThingie from query-params"
+        (ok thingie))
+
+      (PUT* "/echo" []
+        :return   [{:hot Boolean}]
+        :body     [body [{:hot Boolean}]]
+        (ok body))
 
       (POST* "/echo" []
         :return   Thingie
         :body     [thingie Thingie]
-        :summary  "echos a thingie from json-body"
-        (ok thingie))))) ;; here be coerced thingie
+        :summary  "echos a Thingie from json-body"
+        (ok thingie)))))