+++ src/compojure/api/meta.clj
@@ -2,7 +2,7 @@
   (:require [compojure.api.common :refer :all]
             [compojure.core :refer [routes]]
             [ring.util.response :as response]
-            [plumbing.core :refer [letk]]
+            [plumbing.core :refer :all]
             [plumbing.fnk.impl :as fnk-impl]
             [ring.swagger.core :as swagger]
             [ring.swagger.schema :as schema]
@@ -23,20 +23,14 @@
     (fnk-impl/letk-input-schema-and-body-form
       nil (with-meta bind {:schema s/Any}) [] nil)))
 
-(defn resolve-model-var [model]
-  (swagger/resolve-model-var
-    (if (or (set? model) (sequential? model))
-      (first model)
-      model)))
-
 (defn body-coercer-middleware [handler model]
   (fn [request]
     (if-let [response (handler request)]
       (assoc response :body (schema/coerce! model (:body response))))))
 
 (defn src-coerce!
-  "Return source code for coerce! for a schema with coercer type, extracted from a
-   key in a ring request."
+  "Return source code for coerce! for a schema with coercer type,
+   extracted from a key in a ring request."
   [schema key type]
   `(schema/coerce!
      ~schema
@@ -53,8 +47,7 @@
    is consumed form the :parameters map in acc. k = given key, v = value."
   (fn [k v acc] k))
 
-(defmethod restructure-param :default
-  [k v acc]
+(defmethod restructure-param :default [k v acc]
   "By default assoc in the k v to acc"
   (update-in acc [:parameters] assoc k v))
 
@@ -62,79 +55,76 @@
 ;; Default implementation
 ;;
 
-(defmethod restructure-param :return
-  [k model acc]
+(defmethod restructure-param :return [k model acc]
   "Defines a return type and coerced the return value of a body against it."
   (-> acc
-      (update-in [:parameters] assoc k (swagger/resolve-model-vars model))
+      (update-in [:parameters] assoc k (eval model))
       (update-in [:middlewares] conj `(body-coercer-middleware ~model))))
 
-(defmethod restructure-param :body
-  [_ [value model model-meta] acc]
+(defmethod restructure-param :body [_ [value model model-meta] acc]
   "reads body-params into a enchanced let. First parameter is the let symbol,
    second is the Model to coerced! against, third parameter is optional meta-
    data for the model. Examples:
    :body [user User]
-   :body [user User {:key \"value\""
+   :body [user User {:key \"value\"}]"
   (-> acc
       (update-in [:lets] into [value (src-coerce! model :body-params :json)])
       (update-in [:parameters :parameters] conj {:type :body
-                                                 :model (swagger/resolve-model-vars model)
+                                                 :model (eval model)
                                                  :meta model-meta})))
 
-(defmethod restructure-param :query
-  [_ [value model model-meta] acc]
+(defmethod restructure-param :query [_ [value model model-meta] acc]
   "reads query-params into a enchanced let. First parameter is the let symbol,
    second is the Model to coerced! against, third parameter is optional meta-
    data for the model. Examples:
    :query [user User]
-   :query [user User {:key \"value\""
+   :query [user User {:key \"value\"}]"
   (-> acc
       (update-in [:lets] into [value (src-coerce! model :query-params :query)])
       (update-in [:parameters :parameters] conj {:type :query
-                                                 :model (swagger/resolve-model-vars model)
+                                                 :model (eval model)
                                                  :meta model-meta})))
 
-(defmethod restructure-param :body-params
-  [_ body-params acc]
-  "restructures body-params with plumbing letk notation. Generates
-   synthetic defs for the models. Example:
+(defmethod restructure-param :body-params [_ body-params acc]
+  "restructures body-params with plumbing letk notation. Example:
    :body-params [id :- Long name :- String]"
-  (let [schema (strict (fnk-schema body-params))
-        model-name (gensym "body-")
-        _ (eval `(schema/defmodel ~model-name ~schema))
+  (let [schema (eval (strict (fnk-schema body-params)))
         coerced-model (gensym)]
     (-> acc
         (update-in [:lets] into [coerced-model (src-coerce! schema :body-params :json)])
-        (update-in [:parameters :parameters] conj {:type :body :model (eval `(var ~model-name))})
+        (update-in [:parameters :parameters] conj {:type :body :model schema})
         (update-in [:letks] into [body-params coerced-model]))))
 
-(defmethod restructure-param :query-params
-  [_ query-params acc]
-  "restructures query-params with plumbing letk notation. Generates
-   synthetic defs for the models. Example:
+(defmethod restructure-param :query-params [_ query-params acc]
+  "restructures query-params with plumbing letk notation. Example:
    :query-params [id :- Long name :- String]"
-  (let [schema (fnk-schema query-params)
+  (let [schema (eval (fnk-schema query-params))
         coerced-model (gensym)]
     (-> acc
         (update-in [:lets] into [coerced-model (src-coerce! schema :query-params :query)])
-        (update-in [:parameters :parameters] conj {:type :query :model (eval schema)})
+        (update-in [:parameters :parameters] conj {:type :query :model schema})
         (update-in [:letks] into [query-params coerced-model]))))
 
-(defmethod restructure-param :path-params
-  [_ path-params acc]
-  "restructures path-params by plumbing letk notation. Generates
-   synthetic defs for the models. Example:
+#_(defmethod compojure.api.meta/restructure-param :header-params [_ header-params acc]
+  "restructures headers with plumbing letk notation. Example:
+   :query-params [id :- Long name :- String]"
+  (let [schema (eval (fnk-schema header-params))
+        coerced-model (gensym)]
+    (-> acc
+        (update-in [:lets] into [coerced-model (src-coerce! schema :headers :query)])
+        (update-in [:letks] into [header-params coerced-model]))))
+
+(defmethod restructure-param :path-params [_ path-params acc]
+  "restructures path-params by plumbing letk notation. Example:
    :path-params [id :- Long name :- String]"
-  (let [schema (fnk-schema path-params)
+  (let [schema (eval (fnk-schema path-params))
         coerced-model (gensym)]
     (-> acc
         (update-in [:lets] into [coerced-model (src-coerce! schema :route-params :query)])
-        (update-in [:parameters :parameters] conj {:type :path :model (eval schema)})
+        (update-in [:parameters :parameters] conj {:type :path :model schema})
         (update-in [:letks] into [path-params coerced-model]))))
 
-(defmethod restructure-param :middlewares
-  [_ middlewares acc]
+(defmethod restructure-param :middlewares [_ middlewares acc]
   "Applies the given vector of middlewares for the route from left to right"
   (assert (and (vector? middlewares) (every? (comp ifn? eval) middlewares)))
   (update-in acc [:middlewares] into (reverse middlewares)))
@@ -177,13 +167,9 @@
                              (restructure-param k v acc)))
                          (map-of lets letks middlewares parameters body)
                          parameters)
-        body `(~method-symbol
-                ~path
-                ~arg-with-request
-                (meta-container ~parameters
-                                (let ~lets
-                                  (letk ~letks
-                                    ~@body))))]
-    (if (empty? middlewares)
-      body
-      `(middlewares [~@middlewares] ~body))))
+        body `(do ~@body)
+        body (if (seq letks) `(letk ~letks ~body) body)
+        body (if (seq lets) `(let ~lets ~body) body)
+        body (if (seq parameters) `(meta-container ~parameters ~body) body)
+        body `(~method-symbol ~path ~arg-with-request ~body)]
+    (if (seq middlewares) `(middlewares [~@middlewares] ~body) body)))